tests/wpt/web-platform-tests/largest-contentful-paint/expanded-image.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Largest Contentful Paint: expanded image bounded by intrinsic size.</title>
<style type="text/css">
  #image_id {
    width: 300px;
    height: 300px;
  }
</style>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  async_test(function (t) {
    if (!window.LargestContentfulPaint) {
      assert_unreached("LargestContentfulPaint is not implemented");
    }
    let beforeRender = performance.now();
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        assert_equals(entry.entryType, 'largest-contentful-paint');
        assert_greater_than_equal(entry.renderTime, beforeRender,
          'The rendering timestamp should occur after script starts running.');
        assert_greater_than_equal(performance.now(), entry.renderTime,
          'The rendering timestamp should occur before the entry is dispatched to the observer.');
        assert_equals(entry.startTime, entry.renderTime, 'startTime should equal renderTime');
        assert_equals(entry.duration, 0);
        // black-rectangle.png is 100 x 50. It occupies 300 x 300 so size will be bounded by the intrinsic size.
        assert_equals(entry.size, 5000);
        assert_equals(entry.id, 'image_id');
        const pathname = window.location.origin + '/images/black-rectangle.png';
        assert_equals(entry.url, pathname);
        assert_equals(entry.element, document.getElementById('image_id'));
      })
    );
    observer.observe({type: 'largest-contentful-paint', buffered: true});
  }, 'Largest Contentful Paint: |size| attribute is bounded by intrinsic size.');
</script>
<img src='/images/black-rectangle.png' id='image_id'/>
</body>
